+++ source/locationHelper.py
@@ -1,7 +1,7 @@
 # A part of NonVisual Desktop Access (NVDA)
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
-# Copyright (C) 2017-2024 NV Access Limited, Babbage B.V.
+# Copyright (C) 2017-2026 NV Access Limited, Babbage B.V.
 
 """Classes and helper functions for working with rectangles and coordinates."""
 
@@ -293,6 +293,18 @@ class _RectMixin:
 			return RectLTWH(left, top, self.width, self.height)
 		return RectLTRB(left, top, left + self.width, top + self.height)
 
+	def toLTRB(self) -> "RectLTRB":
+		if isinstance(self, RectLTWH):
+			return RectLTRB(self.left, self.top, self.right, self.bottom)
+		assert isinstance(self, RectLTRB)
+		return self
+
+	def toLTWH(self) -> "RectLTWH":
+		if isinstance(self, RectLTRB):
+			return RectLTWH(self.left, self.top, self.width, self.height)
+		assert isinstance(self, RectLTWH)
+		return self
+
 	@property
 	def topLeft(self):
 		return Point(self.left, self.top)
@@ -431,9 +443,6 @@ class RectLTWH(_RectMixin, _RectLTWH):
 	def bottom(self) -> int:
 		return self.top + self.height
 
-	def toLTRB(self) -> "RectLTRB":
-		return RectLTRB(self.left, self.top, self.right, self.bottom)
-
 
 class _RectLTRB(NamedTuple):
 	left: int
@@ -462,9 +471,6 @@ class RectLTRB(_RectMixin, _RectLTRB):
 	@property
 	def height(self) -> int:
 		return self.bottom - self.top
-
-	def toLTWH(self) -> "RectLTWH":
-		return RectLTWH(self.left, self.top, self.width, self.height)
 
 
 #: Classes which support conversion to locationHelper Points using their x and y properties.