+++ tests/unit/test_locationHelper.py
@@ -1,7 +1,7 @@
 # A part of NonVisual Desktop Access (NVDA)
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
-# Copyright (C) 2017-2021 NV Access Limited, Babbage B.V., Łukasz Golonka
+# Copyright (C) 2017-2026 NV Access Limited, Babbage B.V., Łukasz Golonka, hwf1324
 
 """Unit tests for the locationHelper module."""
 
@@ -165,6 +165,24 @@ class TestRectUtilities(unittest.TestCase):
 			RectLTWH(left=10, top=10, width=20, height=20),
 			RectLTWH.fromFloatCollection(10.0, 10.0, 20.0, 20.0),
 		)
+
+	def test_toLTRB(self):
+		left, top, width, height = 10, 10, 20, 20
+		rectLTWH = RectLTWH(left, top, width, height)
+		rectLTRB = RectLTRB(left, top, left + width, top + height)
+		self.assertEqual(rectLTWH.toLTRB(), rectLTRB)
+		self.assertIsInstance(rectLTWH.toLTRB(), RectLTRB)
+		self.assertEqual(rectLTRB.toLTRB(), rectLTRB)
+		self.assertIsInstance(rectLTRB.toLTRB(), RectLTRB)
+
+	def test_toLTWH(self):
+		left, top, right, bottom = 10, 10, 20, 20
+		rectLTRB = RectLTRB(left, top, right, bottom)
+		rectLTWH = RectLTWH(left, top, right - left, bottom - top)
+		self.assertEqual(rectLTRB.toLTWH(), rectLTWH)
+		self.assertIsInstance(rectLTRB.toLTWH(), RectLTWH)
+		self.assertEqual(rectLTWH.toLTWH(), rectLTWH)
+		self.assertIsInstance(rectLTWH.toLTWH(), RectLTWH)
 
 	def test_valueErrorForUnsuportedInput(self):
 		self.assertRaises(ValueError, RectLTRB, left=10, top=10, right=9, bottom=9)